pharmasight/backend/requirements.txt
# PharmaSight Backend Dependencies

# FastAPI & Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.25  # Updated for Python 3.13 compatibility
psycopg2-binary==2.9.9
alembic==1.12.1

# Environment & Config
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator>=2.0.0  # Required for EmailStr validation

# Authentication (Supabase)
supabase==2.0.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0

# Rate limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Excel Import
pandas>=2.0.0
openpyxl>=3.1.0

# PDF (stock take recording template, PO with logo/stamp/signature)
reportlab>=4.0.0
Pillow>=10.0.0
# KRA eTIMS receipt QR on PDF
qrcode[pil]>=7.4.2
requests>=2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

